:sparkles: Namespaces: Decorator
